# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WhyNoIPv6 API
  description: >
    The public read API of whynoipv6.com (07-api.md). Versionless root,
    snake_case wire, `{items,page,meta}` / `{points,meta}` envelopes, keyset
    cursor pagination, RFC 9457 problem+json errors. The only write path is
    the anonymous async live check (`POST /check`). Data is licensed
    CC-BY-NC-4.0.
  version: "1.0.0"
  contact:
    url: https://whynoipv6.com
  license:
    name: CC-BY-NC-4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: https://api.whynoipv6.com
# The filter composition every tier/scoped list shares with /domains
# (07 §4.4). YAML anchor — expanded per path at parse time.
...
paths:
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: The general filterable domain leaderboard
      parameters:
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/country"
        - $ref: "#/components/parameters/asn"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimBase"
        - $ref: "#/components/parameters/dimWWW"
        - $ref: "#/components/parameters/dimNS"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/dimConn"
        - $ref: "#/components/parameters/dimResources"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /domains/{host}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Domain detail with masked informational dimensions
      parameters:
        - $ref: "#/components/parameters/host"
        - name: include
          in: query
          description: "`evidence` attaches the latest scan_detail payload"
          schema: { type: string, enum: [evidence] }
      responses:
        "200":
          description: The §4.3 detail representation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DomainDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /heroes:
    get:
      operationId: listHeroes
      tags: [tiers]
      summary: "Tier preset: class=hero"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /sinners:
    get:
      operationId: listSinners
      tags: [tiers]
      summary: "Tier preset: class=sinner"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /saints:
    get:
      operationId: listSaints
      tags: [tiers]
      summary: "Tier preset: saint=true"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
...
  /countries/{code}/domains:
    get:
      operationId: listCountryDomains
      tags: [countries]
      summary: The country-scoped leaderboard (≡ /domains?country=)
      parameters:
        - $ref: "#/components/parameters/countryCode"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
...
  /asns/{number}/domains:
    get:
      operationId: listASNDomains
      tags: [networks]
      summary: The network-scoped leaderboard (≡ /domains?asn=)
      parameters:
        - $ref: "#/components/parameters/asnNumber"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
...
  /providers/{id}/domains:
    get:
      operationId: listProviderDomains
      tags: [providers]
      summary: The provider-scoped leaderboard (the path form is its own indexed scope)
      parameters:
        - $ref: "#/components/parameters/providerID"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
...
  /campaigns/{uuid}/domains:
    get:
      operationId: listCampaignDomains
      tags: [campaigns]
      summary: Campaign members, host-ordered (rank-NULL rows visible)
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
    host:
      name: host
      in: path
      required: true
      description: eTLD+1 or subdomain; canonicalized before lookup.
      schema: { type: string, example: example.com }
...
    cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from `page.next_cursor`.
      schema: { type: string }
    limit:
      name: limit
      in: query
      description: Page size; default 50, cap 200 (CSV raises the cap to `export.csv_max_rows`).
      schema: { type: integer, minimum: 1, default: 50 }
    afterRank:
      name: after_rank
      in: query
      description: Stateless deep link; rank-ordered views only.
      schema: { type: integer, minimum: 0 }
    sort:
      name: sort
      in: query
      schema: { type: string, enum: [rank, "-rank", host], default: rank }
    class:
      name: class
      in: query
      schema: { $ref: "#/components/schemas/Classification" }
    saint:
      name: saint
      in: query
      description: The only accepted value is true (saint ⊂ hero).
      schema: { type: boolean }
    country:
      name: country
      in: query
      x-go-name: CountryParam
      description: ISO code scope; composes with class + rank order.
      schema: { type: string }
    asn:
      name: asn
      in: query
      description: AS-number scope.
      schema: { type: integer, format: int64 }
    tld:
      name: tld
      in: query
      description: eTLD-suffix pivot; scope-required.
      schema: { type: string }
    provider:
      name: provider
      in: query
      x-go-name: ProviderParam
      description: dns_provider id pivot; scope-required as a query filter.
      schema: { type: integer, format: int64 }
    hosting:
      name: hosting
      in: query
      description: Hosting/CDN text-tag pivot; scope-required.
      schema: { type: string }
    flag:
      name: flag
      in: query
      description: class_flags residual; scope-required.
      schema: { $ref: "#/components/schemas/ClassFlag" }
    dimBase:
      name: base
      in: query
      description: Per-dimension confirmed-status residual; scope-required.
      schema: { $ref: "#/components/schemas/IPv6Status" }
    dimWWW:
      name: www
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
    dimNS:
      name: ns
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
    dimMX:
      name: mx
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
    dimConn:
      name: conn
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
    dimResources:
      name: resources
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
    rankMin:
      name: rank_min
      in: query
      schema: { type: integer, minimum: 0 }
    rankMax:
      name: rank_max
      in: query
      schema: { type: integer, minimum: 0 }
    q:
      name: q
      in: query
      description: Substring search; forces the host ordering.
      schema: { type: string }
    fields:
      name: fields
      in: query
      description: Sparse fieldset (comma-separated top-level keys).
      schema: { type: string }
    format:
      name: format
      in: query
      schema: { type: string, enum: [json, csv], default: json }
...
  responses:
    DomainList:
      description: The `{items,page,meta}` envelope of §4.2 summary rows.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DomainList" }
        text/csv:
          schema: { type: string }
...
    NotFound:
      description: Unknown or uncanonicalizable resource key.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidParameter:
      description: Malformed cursor, bad format, malformed badge host, bad window.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ScopeRequired:
      description: >
        A valid residual filter that needs an indexed companion scope
        (class, country, or asn), or more than one residual was stacked.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    DomainDetail:
      type: object
      required: [host, rank, kind, parent, classification, class_flags, saint, ipv6_only, status, informational, tld, country, asn, dns_provider, hosting_provider, subdomain_count, disabled, last_checked_at, created_at, meta]
      properties:
        host: { type: string }
        rank: { type: integer, nullable: true }
        kind: { type: string, enum: [apex, subdomain] }
        parent: { type: string, nullable: true }
        classification: { $ref: "#/components/schemas/Classification" }
        class_flags:
          type: array
          items: { $ref: "#/components/schemas/ClassFlag" }
        saint: { type: boolean }
        ipv6_only:
          allOf: [{ $ref: "#/components/schemas/IPv6Status" }]
          nullable: true
          description: Derived conn+resources fold — whether the site presents the same over an IPv6-only connection. Null until both dimensions are confirmed.
        status: { $ref: "#/components/schemas/StatusBlock" }
        informational: { $ref: "#/components/schemas/Informational" }
        tld: { type: string, nullable: true }
        country: { $ref: "#/components/schemas/CountryRef" }
        asn: { $ref: "#/components/schemas/ASNRef" }
        dns_provider:
          allOf: [{ $ref: "#/components/schemas/ProviderRef" }]
          nullable: true
        hosting_provider: { type: string, nullable: true }
        subdomain_count: { type: integer, format: int64 }
        disabled: { type: boolean }
        last_checked_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        evidence:
          type: object
          description: Latest scan_detail payload (the §5.1.3 result shape); only with `?include=evidence`.
          additionalProperties: true
        meta: { $ref: "#/components/schemas/DetailMeta" }
...